did some code clean-up

+++ src/com/git/ifly6/communique/ngui/CommuniqueTextDialog.java
@@ -51,7 +51,9 @@ class CommuniqueTextDialog extends JDialog {
 		this.setMinimumSize(new Dimension(300, 350));
 
 		Dimension sSize = Toolkit.getDefaultToolkit().getScreenSize();
-		this.setLocation(Math.round(sSize.width / 2 - width / 2), Math.round(sSize.height / 2 - height / 2));
+		this.setLocation(
+				Math.round((float) sSize.width / 2 - (float) width / 2),
+				Math.round((float) sSize.height / 2 - (float) height / 2));
 
 		JPanel panel = new JPanel();
 		panel.setLayout(new BorderLayout(5, 5));
